trying to clean the setup in order to have it modularized.

+++ lib/DHTLib/dht.cpp
@@ -113,7 +113,7 @@ int dht::_readSensor(uint8_t pin, uint8_t wakeupDelay)
     // direct port read is about 3x faster
 	uint8_t bit = digitalPinToBitMask(pin);
 	uint8_t port = digitalPinToPort(pin);
-    volatile uint8_t *PIR = portInputRegister(port);
+    volatile uint32_t *PIR = portInputRegister(port);
 
     // EMPTY BUFFER
     for (uint8_t i = 0; i < 5; i++) bits[i] = 0;
